neue Berechtigungen

+++ vilesci/api/studiengang/bewerbungstermin/delete_bewerbungstermin.php
@@ -6,11 +6,19 @@ require_once('../../../../../../include/benutzerberechtigung.class.php');
 require_once('../../../../include/bewerbungstermin.class.php');
 require_once('../../functions.php');
 
+$uid = get_uid();
+$berechtigung = new benutzerberechtigung();
+$berechtigung->getBerechtigungen($uid);
+if(!$berechtigung->isBerechtigt("stgv/deleteBewerbungsfrist",null,"suid"))
+{
+    $error = array("message"=>"Sie haben nicht die Berechtigung um Bewerbungsfristen zu löschen.", "detail"=>"stgv/deleteBewerbungsfrist");
+    returnAJAX(FALSE, $error);
+}
+
 $data = filter_input_array(INPUT_POST, array("data"=> array('flags'=> FILTER_REQUIRE_ARRAY)));
 $data = (Object) $data["data"];
 $bewerbungstermin = mapDataToBewerbungstermin($data);
 $bewerbungstermin_id = $bewerbungstermin->bewerbungstermin_id;
- 
 
 if($bewerbungstermin->delete($bewerbungstermin_id))
 {
