+++ vilesci/api/studiengang/doktorat/delete_doktorat.php
@@ -6,11 +6,19 @@ require_once('../../../../../../include/benutzerberechtigung.class.php');
 require_once('../../../../include/doktorat.class.php');
 require_once('../../functions.php');
 
+$uid = get_uid();
+$berechtigung = new benutzerberechtigung();
+$berechtigung->getBerechtigungen($uid);
+if(!$berechtigung->isBerechtigt("stgv/deleteDoktorat",null,"suid"))
+{
+    $error = array("message"=>"Sie haben nicht die Berechtigung um Doktoratsstudienverordnungen zu löschen.", "detail"=>"stgv/deleteDoktorat");
+    returnAJAX(FALSE, $error);
+}
+
 $data = filter_input_array(INPUT_POST, array("data"=> array('flags'=> FILTER_REQUIRE_ARRAY)));
 $data = (Object) $data["data"];
 $doktorat = mapDataToDoktorat($data);
 $doktorat_id = $doktorat->doktorat_id;
- 
 
 if($doktorat->delete($doktorat_id))
 {
